# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Target:
    description: Target entity that needs to be completed
    properties:
      complete:
        type: boolean
      country:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
    type: object
...
  target.UpdateNotesRequest:
    properties:
      notes:
        example: Target usually visits gym at 6 PM
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing spy cats, missions, and targets with JWT authentication
  title: Spy Cat Agency API
  version: "1.0"
paths:
...
  /missions/{id}/targets:
    post:
      consumes:
      - application/json
      description: Add a new target to an existing mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Target'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Target'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add target to mission
      tags:
      - targets
  /missions/{id}/targets/{tid}:
    delete:
      description: Delete a target from mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target ID
        in: path
        name: tid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete target
      tags:
      - targets
  /missions/{id}/targets/{tid}/complete:
    post:
      description: Mark a specific target as completed
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target ID
        in: path
        name: tid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Target'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark target as complete
      tags:
      - targets
  /missions/{id}/targets/{tid}/notes:
    patch:
      consumes:
      - application/json
      description: Update notes for a specific target
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target ID
        in: path
        name: tid
        required: true
        type: integer
      - description: Target notes
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/target.UpdateNotesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Target'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update target notes
      tags:
      - targets
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"